split page in sub components
